refactor: 게시글에 유저 정보 및 권한 추가 (#68)
