--- src/main/java/com/fastcampus/board/board/model/Role.java
@@ -1,5 +1,0 @@
-package com.fastcampus.board.board.model;
-
-public enum Role {
-    SESAC, EXCELLENT;
-}
